refactor: Rename `…package/compile/events/msk/getMskClusterNameToken.js`
