+++ lib/plugins/aws/package/compile/events/msk/get-msk-cluster-name-token.js
