+++ lib/plugins/aws/package/compile/events/msk/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const getMskClusterNameToken = require('./getMskClusterNameToken');
+const getMskClusterNameToken = require('./get-msk-cluster-name-token');
 
 class AwsCompileMSKEvents {
   constructor(serverless) {
